refac: util 분리 - cart
