+++ refactor-record/cart-utils-extraction.md
@@ -1,0 +1,265 @@
+# 장바구니 유틸 함수 분리
+
+## 🚨 코드 스멜
+
+### **비즈니스 로직과 DOM 조작의 강결합**
+
+```javascript
+// 장바구니 추가 로직이 DOM과 강하게 결합
+function handleAddToCart() {
+  const sel = document.getElementById('product-select');
+  const selItem = sel.value;
+  const itemToAdd = findProductById(selItem);
+  
+  // 비즈니스 로직 + DOM 조작이 뒤섞임
+  if (itemToAdd && itemToAdd.quantity > 0) {
+    let item = document.getElementById(itemToAdd.id);
+    if (item) {
+      const qtyElem = item.querySelector('.quantity-number');
+      const currentQty = getQuantityFromElement(qtyElem);
+      const newQty = currentQty + 1;
+      if (newQty <= itemToAdd.quantity + currentQty) {
+        qtyElem.textContent = newQty; // DOM 직접 조작
+        itemToAdd.quantity--; // 비즈니스 로직
+      }
+    }
+  }
+}
+```
+
+- 재고 확인, 수량 계산 등 비즈니스 로직이 DOM 조작과 섞여있음
+- 테스트하기 어려운 구조 (DOM 없이는 로직 검증 불가)
+- 리액트 등 다른 프레임워크에서 재사용 불가
+
+### **중복된 장바구니 업데이트 패턴**
+
+```javascript
+// handleAddToCart에서
+const cartDisp = document.getElementById('cart-items');
+const cartItems = cartDisp.children;
+const cartData = calculateCartTotals(cartItems, products);
+const discountData = calculateDiscounts(cartData.subTot, cartData.totalAmt, cartData.itemCnt);
+const finalTotal = updateOrderSummary({ ...cartData, cartItems }, discountData);
+handleStockInfoUpdate();
+doRenderBonusPoints(finalTotal, cartData.itemCnt);
+
+// handleQuantityChange에서 (동일한 패턴 반복)
+const cartDisp = document.getElementById('cart-items');
+const cartItems = cartDisp.children;
+const cartData = calculateCartTotals(cartItems, products);
+// ... 7줄이 계속 반복됨
+
+// handleRemoveItem에서도 (또 동일한 패턴)
+const cartDisp = document.getElementById('cart-items');
+// ... 똑같은 코드가 또 반복됨
+```
+
+- 3개 함수에서 동일한 7줄 코드가 중복
+- 수정 시 3곳을 모두 찾아서 변경해야 함
+- 실수로 한 곳만 수정하면 버그 발생 위험
+
+### **함수 인자가 너무 많음**
+
+```javascript
+// 인자 4개나 받는 함수
+function calculateCartTotals(cartItems, findProductById, getQuantityFromElement, getProductDiscount) {
+  // ...
+}
+
+// 호출할 때마다 4개 인자를 매번 전달
+const cartData = calculateCartTotals(cartItems, findProductById, getQuantityFromElement, getProductDiscount);
+```
+
+- 함수 호출 코드가 길어짐
+- 인자 순서를 헷갈리기 쉬움
+- 의존성 주입이 명시적이지 않음
+
+## ⚠️ 문제 시나리오
+
+### **비즈니스 로직 테스트 불가**
+
+```javascript
+// 이 함수는 DOM 없이 테스트할 수 없음
+function handleAddToCart() {
+  const sel = document.getElementById('product-select'); // DOM 의존
+  // 재고 확인 로직
+  // 수량 계산 로직
+  // 오류 처리 로직
+  // 모든 로직이 DOM과 섞여있어서 단위 테스트 불가
+}
+```
+
+### **리액트 마이그레이션 시 로직 재작성 필요**
+
+```javascript
+// 바닐라 JS (현재)
+function handleAddToCart() {
+  const sel = document.getElementById('product-select');
+  // DOM 조작 + 비즈니스 로직
+}
+
+// 리액트로 옮길 때 (비즈니스 로직도 다시 작성해야 함)
+function useCart() {
+  const addToCart = (productId) => {
+    // 재고 확인 로직을 처음부터 다시 작성
+    // 수량 계산 로직을 처음부터 다시 작성
+    // 오류 처리 로직을 처음부터 다시 작성
+  };
+}
+```
+
+### **코드 중복으로 인한 버그**
+
+```javascript
+// handleAddToCart에서는 재고 확인을 이렇게 했는데
+if (newQty <= itemToAdd.quantity + currentQty) {
+  // ...
+}
+
+// handleQuantityChange에서는 이렇게 해서 로직이 다름
+if (newQty > 0 && newQty <= prod.quantity + currentQty) {
+  // ...
+}
+```
+
+## ✅ 개선 후 효과
+
+### **비즈니스 로직과 DOM 조작 완전 분리**
+
+```javascript
+// utils/cartUtils.js - 순수한 비즈니스 로직
+export function addItemToCart(productId, currentCartQuantity = 0) {
+  const product = findProductById(productId);
+  
+  if (!product) {
+    return { success: false, error: '상품을 찾을 수 없습니다.' };
+  }
+  
+  if (product.quantity <= 0) {
+    return { success: false, error: '품절된 상품입니다.' };
+  }
+  
+  const newCartQuantity = currentCartQuantity + 1;
+  
+  if (newCartQuantity > product.quantity + currentCartQuantity) {
+    return { success: false, error: '재고가 부족합니다.' };
+  }
+  
+  return {
+    success: true,
+    newCartQuantity,
+    isNewItem: currentCartQuantity === 0,
+    product,
+  };
+}
+
+// main.basic.js - DOM 조작만
+function handleAddToCart() {
+  // 현재 상태 확인
+  const result = addItemToCart(selItem, currentCartQuantity);
+  
+  if (!result.success) {
+    alert(result.error);
+    return;
+  }
+  
+  // DOM 업데이트만
+  if (existingItem) {
+    qtyElem.textContent = result.newCartQuantity;
+  } else {
+    cartDisp.addItem(result.product);
+  }
+}
+```
+
+### **함수 인자 개수 대폭 감소**
+
+```javascript
+// Before: 4개 인자
+calculateCartTotals(cartItems, findProductById, getQuantityFromElement, getProductDiscount)
+
+// After: 1개 인자 (의존성을 내부에서 직접 import)
+calculateCartTotals(cartItems)
+```
+
+### **일관된 오류 처리 패턴**
+
+```javascript
+// 모든 카트 유틸 함수가 동일한 응답 형식
+const result = addItemToCart(productId, currentQty);
+if (!result.success) {
+  alert(result.error); // 일관된 오류 처리
+  return;
+}
+
+const result2 = updateCartItemQuantity(productId, change, currentQty);
+if (!result2.success) {
+  alert(result2.error); // 같은 패턴
+  return;
+}
+```
+
+### **전역 데이터 중앙화**
+
+```javascript
+// data/products.js - 상품 데이터 중앙 관리
+export let products = [...];
+export function findProductById(productId) {
+  return products.find((product) => product.id === productId);
+}
+
+// utils에서 직접 import하여 사용
+import { findProductById } from '../data/products.js';
+```
+
+## 🎯 리액트 마이그레이션 준비 완료
+
+### **순수 함수들을 그대로 활용 가능**
+
+```javascript
+// React에서 바로 사용 가능
+import { addItemToCart, updateCartItemQuantity, removeCartItem } from './utils/cartUtils.js';
+import { products } from './data/products.js';
+
+function useCart() {
+  const [cartItems, setCartItems] = useState([]);
+  
+  const addItem = (productId) => {
+    const currentQty = cartItems.find(item => item.id === productId)?.quantity || 0;
+    const result = addItemToCart(productId, currentQty);
+    
+    if (result.success) {
+      // 상태 업데이트 (DOM 조작 대신)
+      setCartItems(prev => [...prev, { id: productId, quantity: result.newCartQuantity }]);
+    }
+  };
+  
+  return { cartItems, addItem };
+}
+```
+
+### **테스트 작성 용이**
+
+```javascript
+// 이제 DOM 없이도 비즈니스 로직 테스트 가능
+import { addItemToCart } from '../utils/cartUtils.js';
+
+test('재고 부족 시 오류 반환', () => {
+  const result = addItemToCart('product1', 50); // 재고보다 많이 요청
+  expect(result.success).toBe(false);
+  expect(result.error).toBe('재고가 부족합니다.');
+});
+
+test('정상 추가 시 성공 반환', () => {
+  const result = addItemToCart('product1', 0);
+  expect(result.success).toBe(true);
+  expect(result.newCartQuantity).toBe(1);
+});
+```
+
+## 📊 개선 지표
+
+- **코드 중복**: 7줄 × 3곳 = 21줄 → 0줄 (100% 제거)
+- **함수 인자**: 4개 → 1~2개 (50~75% 감소)
+- **테스트 가능**: 0% → 100% (비즈니스 로직 완전 분리)
+- **재사용성**: 바닐라 JS 전용 → 프레임워크 무관 (범용성 확보)
